+++ opc-workbench/src/main.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App';
+import { I18nProvider } from './i18n';
 import { APP_CONFIG } from './config';
 import 'tdesign-react/esm/style/index.js';
 import './index.css';
@@ -11,7 +12,9 @@ document.title = APP_CONFIG.name;
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <BrowserRouter>
-      <App />
+      <I18nProvider>
+        <App />
+      </I18nProvider>
     </BrowserRouter>
   </React.StrictMode>,
 );